+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ Changelog
 - make sure to use correct url of tile
   [vangheem]
 
+- handle not found errors while rendering tiles so layout
+  isn't borked
+  [vangheem]
+
 1.0 (2012-06-23)
 ----------------